+++ package.json
@@ -5,6 +5,8 @@
 	"dependencies": {
 		"@emotion/react": "^11.9.0",
 		"@emotion/styled": "^11.8.1",
+		"@linaria/core": "^3.0.0-beta.22",
+		"@linaria/react": "^3.0.0-beta.22",
 		"apexcharts": "^3.33.0",
 		"axios": "^0.27.2",
 		"axios-mock-adapter": "^1.21.1",
@@ -45,7 +47,10 @@
 		]
 	},
 	"devDependencies": {
+		"@linaria/babel": "^3.0.0-beta.0",
+		"@linaria/webpack-loader": "^3.0.0-beta.22",
 		"babel-eslint": "8",
+		"craco-linaria3": "^1.0.1",
 		"eslint": "4.x",
 		"eslint-config-prettier": "^8.3.0",
 		"eslint-plugin-prettier": "^4.0.0",
