--- src/components/App.css
@@ -1,7 +1,0 @@
-* {
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-/*# sourceMappingURL=App.css.map */
