--- src/components/Company/css/DashBoard.css
@@ -1,131 +1,0 @@
-.dashboard {
-  width: 30%;
-  height: 100%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-  overflow-y: auto;
-  padding-right: 15px;
-}
-
-@media screen and (max-width: 820px) {
-  .dashboard {
-    width: 100%;
-    height: 100%;
-    display: -ms-grid;
-    display: grid;
-    -ms-grid-rows: 350px 350px 350px;
-        grid-template-rows: 350px 350px 350px;
-    -ms-grid-columns: 1fr 1fr;
-        grid-template-columns: 1fr 1fr;
-    row-gap: 5px;
-    -webkit-column-gap: 5px;
-            column-gap: 5px;
-    margin-top: 10px;
-    padding: 0px;
-    overflow-y: visible;
-  }
-  .dashboard .profile-card {
-    -ms-grid-row: 1;
-    -ms-grid-row-span: 2;
-    grid-row: 1 / 3;
-    -ms-grid-column: 1;
-    -ms-grid-column-span: 1;
-    grid-column: 1 / 2;
-  }
-  .dashboard .news-card {
-    -ms-grid-row: 1;
-    -ms-grid-row-span: 1;
-    grid-row: 1 / 2;
-    -ms-grid-column: 2;
-    -ms-grid-column-span: 1;
-    grid-column: 2 / 3;
-  }
-  .dashboard .financials-card {
-    -ms-grid-row: 2;
-    -ms-grid-row-span: 1;
-    grid-row: 2 / 3;
-    -ms-grid-column: 2;
-    -ms-grid-column-span: 1;
-    grid-column: 2 / 3;
-  }
-  .dashboard .earning-card {
-    -ms-grid-row: 3;
-    -ms-grid-row-span: 1;
-    grid-row: 3 / 4;
-    -ms-grid-column: 2;
-    -ms-grid-column-span: 1;
-    grid-column: 2 / 3;
-  }
-  .dashboard .recommend-card {
-    -ms-grid-row: 3;
-    -ms-grid-row-span: 1;
-    grid-row: 3 / 4;
-    -ms-grid-column: 1;
-    -ms-grid-column-span: 1;
-    grid-column: 1 / 2;
-  }
-}
-
-@media screen and (max-width: 480px) {
-  .dashboard {
-    width: 100%;
-    height: 100%;
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-        -ms-flex-direction: column;
-            flex-direction: column;
-    margin-top: 10px;
-  }
-}
-
-.dashboard .card {
-  background-color: white;
-  margin-bottom: 10px;
-  padding: 0.5em;
-}
-
-@media screen and (max-width: 820px) {
-  .dashboard .card {
-    margin-bottom: 0px;
-  }
-}
-
-@media screen and (max-width: 480px) {
-  .dashboard .card:first-child {
-    margin-right: 0px;
-  }
-}
-
-.dashboard .card .card-title {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: horizontal;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: row;
-          flex-direction: row;
-  -webkit-box-pack: justify;
-      -ms-flex-pack: justify;
-          justify-content: space-between;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  font-size: 1.5em;
-  font-weight: bold;
-  padding: 10px;
-}
-
-@media screen and (max-width: 820px) {
-  .dashboard .card .card-title {
-    font-size: 2.5em;
-  }
-}
-/*# sourceMappingURL=DashBoard.css.map */
