--- src/components/Company/css/Profile.css
@@ -1,93 +1,0 @@
-.profile-wrapper {
-  padding: 15px;
-}
-
-.profile-wrapper .profile-title {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-.profile-wrapper .profile-title .profile-logo {
-  width: 60px;
-  padding: 10px;
-  background-color: white;
-  border-radius: 100%;
-  border: 2px solid #05325c;
-  margin-right: 15px;
-}
-
-.profile-wrapper .profile-title .profile-name {
-  font-size: 1.7em;
-  font-weight: bold;
-}
-
-.profile-wrapper .profile-title .profile-symbol {
-  color: #0084ff;
-  margin-top: 5px;
-}
-
-.profile-wrapper .profile-table {
-  width: 100%;
-  margin-top: 15px;
-}
-
-.profile-wrapper .profile-table th {
-  text-align: left;
-}
-
-.profile-wrapper .profile-summary {
-  margin-top: 15px;
-}
-
-.profile-wrapper .fold {
-  height: 200px;
-  overflow: hidden;
-}
-
-@media screen and (max-width: 820px) {
-  .profile-wrapper .fold {
-    height: auto;
-  }
-}
-
-@media screen and (max-width: 480px) {
-  .profile-wrapper .fold {
-    height: 200px;
-  }
-}
-
-.profile-wrapper .dashboard-profile-more-btn {
-  width: 100%;
-  height: 40px;
-  margin-top: 15px;
-}
-
-.profile-wrapper .dashboard-profile-more-btn button {
-  width: 100%;
-  height: 100%;
-  border-radius: 5px;
-  background-color: #0084ff;
-  color: white;
-  font-weight: bold;
-}
-
-.profile-wrapper .dashboard-profile-more-btn button:hover {
-  background-color: #0070da;
-}
-
-@media screen and (max-width: 820px) {
-  .profile-wrapper .dashboard-profile-more-btn {
-    display: none;
-  }
-}
-
-@media screen and (max-width: 480px) {
-  .profile-wrapper .dashboard-profile-more-btn {
-    display: block;
-  }
-}
-/*# sourceMappingURL=Profile.css.map */
