--- src/components/SearchList/styles/SearchListStyle.css
@@ -1,145 +1,0 @@
-.wrapper {
-  height: 100%;
-  background-color: #ececec;
-}
-
-.contents {
-  width: 70%;
-  height: 100%;
-  margin: 0 auto;
-  padding-top: 5vh;
-}
-
-.search-result {
-  width: 45%;
-  height: 9vh;
-  background-color: #fff;
-  display: -ms-grid;
-  display: grid;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-.search-result .search-num-text {
-  margin-left: 3vw;
-}
-
-.search-result span {
-  display: inline;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  font-size: 2rem;
-  font-weight: bold;
-}
-
-.search-result .search-num {
-  color: #ff0000;
-}
-
-.search-item {
-  width: 100%;
-  height: 15vh;
-  background-color: #fff;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  margin-top: 5vh;
-}
-
-.search-item:hover {
-  cursor: pointer;
-  background-color: #b6b6b6;
-  -webkit-transition: 0.2s ease-in-out;
-  transition: 0.2s ease-in-out;
-}
-
-.search-item:last-child {
-  margin-bottom: 10vh;
-}
-
-.search-item .company-name {
-  font-size: 2.3rem;
-  font-weight: bold;
-  text-align: center;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  margin-left: 3vw;
-}
-
-.number {
-  width: 80px;
-  height: 80px;
-  background-color: #2e9bff;
-  display: -ms-grid;
-  display: grid;
-  margin: auto 0 auto 2vw;
-}
-
-.number h1 {
-  color: #fff;
-  text-align: center;
-  vertical-align: middle;
-  line-height: 80px;
-  font-size: 3rem;
-}
-
-.stock-numbers {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  margin: auto 6vw auto auto;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-}
-
-.stock-numbers .stock-price {
-  font-size: 2.3rem;
-  font-weight: bold;
-}
-
-.stock-numbers .search-price {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  font-size: 1.25rem;
-  margin-left: 1vw;
-  margin-top: 1vh;
-}
-
-.stock-numbers .search-price span {
-  margin-right: 5px;
-}
-
-.stock-numbers .search-price .up {
-  color: #ff0000;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-}
-
-.stock-numbers .search-price .down {
-  color: #0038ff;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-}
-
-.stock-numbers .search-price .revised-price {
-  margin-right: 1.5vw;
-}
-
-.stock-numbers .search-price .percent {
-  margin-right: 1vw;
-}
-/*# sourceMappingURL=SearchListStyle.css.map */
