+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import 'styles/core.scss';
-import App from 'components/App';
+import App from 'pages/App';
 
 ReactDOM.render(
 	<React.StrictMode>
