+++ src/pages/Company/components/Earning.jsx
