+++ src/pages/Company/components/News.jsx
