+++ src/pages/Company/components/Profile.jsx
