+++ src/pages/SearchList/styles/SearchListStyle.scss
